+++ mllib-dal/src/main/native/oneapi/dal/ColumnAccessorImpl.cpp
@@ -23,7 +23,7 @@
 #include <vector>
 
 #ifdef CPU_GPU_PROFILE
-#include "common.hpp"
+#include "Common.hpp"
 
 #include "com_intel_oneapi_dal_table_ColumnAccessor.h"
 #include "oneapi/dal/table/column_accessor.hpp"
