+++ mllib-dal/src/main/native/oneapi/dal/HomogenTableImpl.cpp
@@ -24,7 +24,7 @@
 #include <mutex>
 
 #ifdef CPU_GPU_PROFILE
-#include "common.hpp"
+#include "Common.hpp"
 
 #include "com_intel_oneapi_dal_table_HomogenTableImpl.h"
 #include "service.h"
